Logic for processing findObjects response in DSL
Added logic for processing the API response for findObjects in the DSL script in a groovy-friendly manner. The script accounts for evalDsl being invoked through ectool and via REST. This special logic was needed as the `findObjects` response type changes based on whether the client requested format is json or xml(default for ectool).